backend/requirements.txt
# FastAPI and server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Document processing
pymupdf>=1.24.0,<2.0.0  # PDF support (pre-built wheels for Windows)
pypdf>=3.0.0  # PDF support for LangChain PyPDFLoader
python-docx>=1.1.0  # DOCX support
python-pptx>=0.6.23  # PPTX support
Pillow>=10.2.0,<13.0.0  # Image processing (10.2.0+ has Python 3.13 support, <13.0 to avoid future conflicts)

# LangChain
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.5

# Vector database
qdrant-client>=1.7.0

# Google AI
google-genai>=0.2.2

# Utilities
tqdm>=4.66.1

